requirements.txt
Flask
SQLAlchemy
Flask-JSON
cerberus
zope.component
zope.interface